Been a day or so since last commit - lots of development bugs and issues to overcome. Working through them though. Now a user can create the form, naming columns, test and title. This goes into the database, and the journal name goes on the recycler view. Now to get the fields from journal structure with the WHERE query but i dont think live data is the way to do this. Come back to this commit if things go badly

+++ app/src/main/java/com/example/diss_cbt_application/JournalFeature/JDatabase/JDTables/JournalObject.java
@@ -19,8 +19,7 @@ public class JournalObject {
 
     /*Table member variables*/
     @PrimaryKey(autoGenerate = true)//id is auto incremented
-    private int id;
-
+    private Long id;
 
     private String journalName;
 
@@ -35,12 +34,12 @@ public class JournalObject {
     }
 
     /*Room uses this to set the id on the object*/
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
     /*To persist these fields room needs getter methods for all of the fields*/
-    public int getId() {
+    public Long getId() {
         return id;
     }
